RPG commands to RPG plugin
