+++ magnet2.py
