+++ plugin_command_simple_fun.py
@@ -216,37 +216,6 @@ def command_status(bot, room, nick, access_level, parameters, message):
     response = '%s is %s and also %s.'%(target, pos, desc)
   return response
 
-def command_roll(bot, room, nick, access_level, parameters, message):
-  p = parameters.find('d')
-  x = parameters.find('+')
-  s = parameters.find(' ')
-  if s != -1: parameters = parameters[:s]
-  number = 0
-  sides = 0
-  add = 0
-  if p == -1: p = len(parameters)
-  try:
-    if x == -1:
-      sides = int(parameters[p+1:])
-    else:
-      sides = int(parameters[p+1:x])
-      add = int(parameters[x+1:])
-    number = int(parameters[:p])
-  except:
-    pass
-  if number < 1: number = 1
-  if number > 10: number = 10
-  if sides < 2: sides = 6
-  if sides > 100: sides = 100
-  if add < -1000: add = -1000
-  if add > 1000: add = 1000
-  rolls = []
-  for i in range(number): rolls.append(random.randint(1, sides))
-  if add: rolls.append(add)
-  output = '%s, %dd%d+%d roll: %s'%(nick, number, sides, add, '+'.join([str(i) for i in rolls]))
-  if len(rolls) > 1: output += '=' + str(sum(rolls))
-  return output
-
 def command_tellfruit(bot, room, nick, access_level, parameters, message):
   if parameters:
     target = parameters
@@ -260,9 +229,6 @@ def command_tellfruit(bot, room, nick, access_level, parameters, message):
   for c in hashname: hashnum += ord(c)
   veg = ("fruit", "vegetable")[hashnum % 2]
   return '%s is a %s.'%(target, veg)
-
-def command_flip(bot, room, nick, access_level, parameters, message):
-  return nick+', '+random.choice(['Heads!', 'Tails!'])
 
 def command_say(bot, room, nick, access_level, parameters, message):
   bot.send_room_message(room, parameters)
@@ -282,8 +248,7 @@ def load(bot):
   bot.add_command('transform', command_transform, LEVEL_GUEST, 'simple_fun')
   bot.add_command('status', command_status, LEVEL_GUEST, 'status')
   bot.add_command('tellfruit', command_tellfruit, LEVEL_GUEST, 'status')
-  bot.add_command('roll', command_roll, LEVEL_GUEST, 'rpg')
-  bot.add_command('flip', command_flip, LEVEL_GUEST, 'rpg')
+  bot.add_command('flip', command_flip, LEVEL_GUEST, 'flip')
   bot.add_command('say', command_say, LEVEL_ADMIN, 'say')
   bot.add_command('act', command_act, LEVEL_ADMIN, 'say')
